Move page components into routes under app/
