+++ src/app/[[...id]]/page.test.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import { MemoryRouter } from "react-router-dom"
 import { screen, render } from "@testing-library/react"
 
-import { ImageDetails } from "./ImageDetails"
+import ImageDetails from "./page"
 
 describe("ImageDetails", () => {
   it("renders image details", () => {
